Delete outdated comparison value in case we do a full load
This prevents problem in case we trigger a full load with TRUNCATE/reCREATE of the table but still have an old comparison value and then crash during that full load. The next run would then see a non-empty table and a comparison value and would do an incremental load which would result in the table missing rows.