Fix inconsistent state after error (#24)
* Delete outdated comparison value in case we do a full load 

This prevents problem in case we trigger a full load with TRUNCATE/reCREATE of the table but still have an old comparison value and then crash during that full load. The next run would then see a non-empty table and a comparison value and would do an incremental load which would result in the table missing rows.

* Delete outdated file dependency hashes before running the command

This prevents problems when the run would usually recreate the state afresh and then update the hashes. But if the recreate fails and the files are reverted to the old state, the DB would not have the new state but the old (matching) hashes would prevent a recreate.