Make it possible to disable output coloring in command line etl runs
In certain cases these show up as plain escape sequences and this gets rid if it.

Also fixes a bug in the help output of data-integration.ui.run: you cannot specifiy multiple pipelines, if you want that, you have to specify the parent as pipeline and the specific pipelines as nodes.