fix exitcode of process issue since python 3.8
The internal exitcode handling of a process changed since python 3.8 (https://github.com/python/cpython/issues/84275) which caused that when executing a pipeline through the UI the exit code from the process is 1 even when using sys.exit(0). When we use the same multiprocessing context this seems to be handled properly in python and we get the correct exit code.