fix web UI execution, exitcode still wrong
The exitcode is sometimes still zero, even I use the multiprocessing context. To get around that issue, we do not interpret the exitcode anymore but rely 100% on the status_queue:
The status queue must return True, otherwise it is counted as unsuccessful. So, in the 'run' method the succeeded status should only be passed to the queue when we are sure that the task execution succeeded.