fix exitcode of process issue since python 3.8 (#87)
* fix exitcode of process issue since python 3.8

The internal exitcode handling of a process changed since python 3.8 (https://github.com/python/cpython/issues/84275) which caused that when executing a pipeline through the UI the exit code from the process is 1 even when using sys.exit(0). When we use the same multiprocessing context this seems to be handled properly in python and we get the correct exit code.

* typo

* fix web UI execution, exitcode still wrong

The exitcode is sometimes still zero, even I use the multiprocessing context. To get around that issue, we do not interpret the exitcode anymore but rely 100% on the status_queue:
The status queue must return True, otherwise it is counted as unsuccessful. So, in the 'run' method the succeeded status should only be passed to the queue when we are sure that the task execution succeeded.

* fix passing multiprocessing_context

* add pipeline failed test